removes unused code
